Setup integration tests to exercise modules
This sets up a test with [test-kitchen](https://rubygems.org/gems/test-kitchen/versions/2.3.4) that spins up resources in a real AWS account and runs assertions to validate they were created as expcted.

This was adapted from https://www.darkraiden.com/blog/test-terraform-with-kitchen-and-awspec/ (with a few differences, such as newer versions of the tools).

This isn't setup to run on CI as it requires AWS credentials, but can be run locally with tools like `aws-okta`, e.g. `aws-okta exec dev-write -- make test`.

This also fixes bugs caught by the new tests:
* fix principal type of the Segment account to be `AWS` not `Service`
* split IAM policies into `assume_role_policy` and a policy attachment